Renames DomainNode.copyMetas -> copy (#83)
Issue #81.

The name "copyMetas" did not reflect what the function was actually
doing.  It does not copy the metas, but rather copies the node while
replacing the metas!  Hence, "copy" is a much better name.

Every generated class now gets up to two "copy" functions: one that
overrides the DomainNode.copy(MetaContainer) function and one that does
not override and is equivalent to the Kotlin data class copy function.
For user-defined types with no elements, the latter ends up with the
same signature as the former, so changes were also included to the
kotlin.ftl file to omit this function in the event that the type has
no elements.